BMP085/BMP180: Added polling for end of conversion bit. Updated example. Removed unnecessary float.

+++ Arduino/BMP085/BMP085.h
@@ -111,7 +111,7 @@ class BMP085 {
         float       getTemperatureC();
         float       getTemperatureF();
         uint32_t    getRawPressure();
-        float       getPressure();
+        int32_t     getPressure();
         float       getAltitude(float pressure, float seaLevelPressure=101325);
 
    private:
